SET UP PYGAME BASE FOR SUDOKU GUI
SudokuGUI class is done and it will now render the sudoku minor grid
lines as well as the major nonet grid lines. This is because the grid
lines seperate the sudoku boxes where the numbers will be places.
Additionally, User I/O will be parsed and stored, but it won't do
anything.